requirements.txt
mem0ai>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fastmcp
fastapi
uvicorn[standard]
google-generativeai
requests
networkx
pyvis
